+++ SharedClasses/Glo.cs
@@ -5,9 +5,9 @@ using System.Xml.Linq;
 using System.Reflection;
 
 // Application version
-[assembly: AssemblyVersion("0.6.5")]
-[assembly: AssemblyFileVersion("0.6.5")]
-[assembly: AssemblyInformationalVersion("0.6.5")]
+[assembly: AssemblyVersion("0.6.6")]
+[assembly: AssemblyFileVersion("0.6.6")]
+[assembly: AssemblyInformationalVersion("0.6.6")]
 
 public static class Glo
 {